requirements.txt
web3==5.31.1
typer[all]==0.7.0
rich==12.6.0
pydantic==1.10.2
tenacity==8.1.0
stakefish-web3-utils==0.0.7
